[:core] [:micro:metadata] Reorganize http event classes
