+++ http/server/src/main/java/io/zero88/qwe/http/server/gateway/GatewayIndexApi.java
@@ -1,6 +1,6 @@
 package io.zero88.qwe.http.server.gateway;
 
-import io.zero88.qwe.http.ActionMethodMapping;
+import io.zero88.qwe.micro.httpevent.ActionMethodMapping;
 import io.zero88.qwe.http.server.HttpServerPlugin;
 import io.zero88.qwe.http.server.rest.api.AbstractRestEventApi;
