+++ http/server/src/main/java/io/zero88/qwe/http/server/rest/api/DynamicEventRestApi.java
@@ -8,8 +8,8 @@ import java.util.stream.Collectors;
 import io.github.zero88.utils.Strings;
 import io.vertx.core.json.JsonObject;
 import io.vertx.servicediscovery.Record;
-import io.zero88.qwe.http.EventMethodDefinition;
-import io.zero88.qwe.http.EventMethodMapping;
+import io.zero88.qwe.micro.httpevent.EventMethodDefinition;
+import io.zero88.qwe.micro.httpevent.EventMethodMapping;
 import io.zero88.qwe.micro.servicetype.EventMessageHttpService;
 
 import lombok.NonNull;
