+++ micro/discovery/src/test/java/io/zero88/qwe/micro/ServiceDiscoveryApiTest.java
@@ -15,7 +15,7 @@ import io.zero88.qwe.JsonHelper;
 import io.zero88.qwe.dto.msg.RequestData;
 import io.zero88.qwe.dto.msg.RequestFilter;
 import io.zero88.qwe.event.EventAction;
-import io.zero88.qwe.http.EventMethodDefinition;
+import io.zero88.qwe.micro.httpevent.EventMethodDefinition;
 import io.zero88.qwe.micro.filter.ByPredicateFactory;
 import io.zero88.qwe.micro.filter.ServiceFilterParam;
 import io.zero88.qwe.micro.filter.ServiceStatusPredicateFactory;
