+++ micro/discovery/src/test/java/io/zero88/qwe/micro/ServiceLocatorTest.java
@@ -18,7 +18,7 @@ import io.zero88.qwe.event.EventAction;
 import io.zero88.qwe.event.EventMessage;
 import io.zero88.qwe.event.EventStatus;
 import io.zero88.qwe.exceptions.ErrorCode;
-import io.zero88.qwe.http.EventMethodDefinition;
+import io.zero88.qwe.micro.httpevent.EventMethodDefinition;
 import io.zero88.qwe.micro.filter.ByPredicateFactory;
 import io.zero88.qwe.micro.filter.ServiceFilterParam;
 import io.zero88.qwe.micro.transfomer.RecordTransformer.ViewType;
