+++ micro/metadata/src/main/java/io/zero88/qwe/micro/httpevent/EventHttpService.java
@@ -1,4 +1,4 @@
-package io.zero88.qwe.http;
+package io.zero88.qwe.micro.httpevent;
 
 import java.util.Set;
