+++ micro/metadata/src/main/java/io/zero88/qwe/micro/servicetype/EventMessageHttpService.java
@@ -6,7 +6,7 @@ import io.github.zero88.utils.Strings;
 import io.vertx.core.json.JsonObject;
 import io.vertx.servicediscovery.Record;
 import io.vertx.servicediscovery.spi.ServiceType;
-import io.zero88.qwe.http.EventMethodDefinition;
+import io.zero88.qwe.micro.httpevent.EventMethodDefinition;
 
 import lombok.NonNull;
