+++ micro/metadata/src/main/java/io/zero88/qwe/micro/transfomer/EventServicePublicView.java
@@ -3,8 +3,8 @@ package io.zero88.qwe.micro.transfomer;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.servicediscovery.Record;
-import io.zero88.qwe.http.EventMethodDefinition;
-import io.zero88.qwe.http.EventMethodMapping;
+import io.zero88.qwe.micro.httpevent.EventMethodDefinition;
+import io.zero88.qwe.micro.httpevent.EventMethodMapping;
 import io.zero88.qwe.micro.servicetype.EventMessageHttpService;
 
 import lombok.NonNull;
