+++ micro/metadata/src/test/java/io/zero88/qwe/micro/httpevent/EventMethodDefinitionTest.java
@@ -1,4 +1,4 @@
-package io.zero88.qwe.http;
+package io.zero88.qwe.micro.httpevent;
 
 import java.util.Arrays;
 import java.util.HashMap;
