+++ micro/metadata/src/test/java/io/zero88/qwe/micro/httpevent/EventMethodMappingTest.java
@@ -1,4 +1,4 @@
-package io.zero88.qwe.http;
+package io.zero88.qwe.micro.httpevent;
 
 import org.json.JSONException;
 import org.junit.jupiter.api.Assertions;
